Updates to notarization work done by SVT - build and sign on macos 10.14.5 (#188)
* Builds, signs and notarizes the app on MacOS 10.14.5

- Sign and Notarize the mac build. https://kilianvalkhof.com/2019/electron/notarizing-your-electron-application/
- Use environment variables to override the default bundleID
- Updates the documentation

* Sets appPath from the context instead of a static string

This makes it easier to change the app name.

* Builds, signs and notarizes the app on MacOS 10.14.5

- Sign and Notarize the mac build. https://kilianvalkhof.com/2019/electron/notarizing-your-electron-application/
- Use environment variables to override the default bundleID
- Updates the documentation

* Sets appPath from the context instead of a static string

This makes it easier to change the app name.

* Don't notarize if not signed


Don't notarize if not signed

* :art: Format using the standard linter

* Updated webpack dev server version and added new mac models

* Added electron-notarize package, allow config writing to be skipped.

* Removed appId env var from build script

* Completed support for electron-notarize

* Don't process.exit if skipping file update

* Renamed scripts/* to reflect that they are run conditionally.

* Use fs.writeFileSync is config updater. Added additional debug logging for build

* Rolled back changes to macmodels database

* Added instructions for `jwt` authentication and ascProvider in notarization process. Fixed typo in build docs.
